% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels-grob.R
\name{add_box}
\alias{add_box}
\title{make frame around text grobs}
\usage{
add_box(grob, width_pt, height_pt, descent_pt, x = unit(0.5, "npc"),
  y = unit(0.5, "npc"), box_hjust = 0.5, box_vjust = 0.5,
  padding = mar(0, 0, 0, 0), margin = mar(0, 0, 0, 0), angle = 0, ...,
  debug = TRUE)
}
\description{
make frame around text grobs
}
