% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box-grob.R
\name{box_grob}
\alias{box_grob}
\title{Draw a box with content}
\usage{
box_grob(content_grob, width = NULL, height = NULL, x = unit(0.5,
  "npc"), y = unit(0.5, "npc"), hjust = 0.5, vjust = 0.5,
  padding = mar(0, 0, 0, 0), margin = mar(0, 0, 0, 0), angle = 0,
  fill = NA, color = NA, ..., debug = FALSE)
}
\arguments{
\item{content_grob}{A grob that represents the contents inside the box.}

\item{width, height}{Width and height of the box grob. If either is set
to \code{NULL} (the default), it is calculated from the content grob.}

\item{x, y}{Unit objects specifying the location of the reference point.}

\item{hjust, vjust}{Numerical values specifying the location of the box
relative to the reference point.}

\item{padding}{Unit vector of 4 specifying the padding, in the order
top, right, bottom, left.}

\item{margin}{Unit vector of 4 specifying the margin, in the order
top, right, bottom, left.}

\item{angle}{Rotation angle (in degrees)}

\item{fill}{Fill color of the box.}

\item{color}{Outline color of the box.}

\item{...}{Not used.}

\item{debug}{Bool indicating whether debug info should be drawn or not.}
}
\description{
This grob draws a box around another grob specifying content.
}
\examples{
library(grid)

g1 <- textGrob("abcd")
g2 <- box_grob(
  g1,
  angle = 40,
  padding = mar(10, 20, 10, 20),
  margin = mar(20, 10, 20, 10),
  hjust = 0,
  vjust = 1,
  debug = TRUE
)
grid.newpage()
grid.draw(g2)
}
