% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-rich-text.R
\name{draw_rich_text}
\alias{draw_rich_text}
\title{Draw rich text}
\usage{
draw_rich_text(contents, x_pt = 50, y_pt = 100, width_pt = 300,
  gp = NULL, newpage = TRUE)
}
\arguments{
\item{contents}{character vector containing html string to draw}

\item{x_pt}{x location, in points}

\item{y_pt}{y location, in points}

\item{width_pt}{width, in points}

\item{gp}{Other graphical parameters for drawing}

\item{newpage}{Bool indicating whether \code{grid.newpage()} should be called first}
}
\description{
Draw rich text
}
\examples{
library(grid)
grid.newpage()
draw_rich_text(
  "This is the first line.<br>This is the second line.<br>And some <b>text in bold.</b>"
)
draw_rich_text(
  "This is some <span style='font-family:\\"Bookman\\"; font-size:20;'>big
   font <span style='font-size:25; color:red'>in red.</span></span> <br>And the next
   line like nothing happened."
)
}
