% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed-box-grob.R
\name{fixed_box_grob}
\alias{fixed_box_grob}
\title{Draw a fixed-sized box with content}
\usage{
fixed_box_grob(content_grob, width_pt = 100, height_pt = 100,
  x = unit(0.5, "npc"), y = unit(0.5, "npc"), hjust = 0.5,
  vjust = 0.5, padding = mar(0, 0, 0, 0), margin = mar(0, 0, 0, 0),
  angle = 0, fill = NA, color = NA, ..., debug = FALSE)
}
\arguments{
\item{content_grob}{A grob that represents the contents inside the box.}

\item{width_pt, height_pt}{Width and height of the box grob, in
points ("pt").}

\item{x, y}{Unit objects specifying the location of the reference point.}

\item{hjust, vjust}{Numerical values specifying the location of the box
relative to the reference point.}

\item{padding}{Unit vector of 4 specifying the padding, in the order
top, right, bottom, left. Must be specified in absolute units (point,
mm, inch, etc.).}

\item{margin}{Unit vector of 4 specifying the margin, in the order
top, right, bottom, left. Must be specified in absolute units (point,
mm, inch, etc.).}

\item{angle}{Rotation angle (in degrees)}

\item{fill}{Fill color of the box.}

\item{color}{Outline color of the box.}

\item{debug}{Bool indicating whether debug info should be drawn or not.}
}
\description{
This grob draws a box around another grob specifying content. All sizes need
to be specified in absolute units ahead of time; this enables much faster
unit calculations.
}
\examples{
library(grid)

g1 <- textGrob("abcd")
g2 <- box_grob(
  g1,
  width_pt = convertWidth(grobWidth(g1), "pt"),
  height_pt = convertHeight(grobHeight(g1), "pt"),
  angle = 40,
  padding = mar(10, 20, 10, 20),
  margin = mar(20, 10, 20, 10),
  hjust = 0,
  vjust = 1,
  debug = TRUE
)
grid.newpage()
grid.draw(g2)
}
