% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob-descent.R
\name{font_details_pt}
\alias{font_details_pt}
\title{Calculates generic font height and descent from given graphical parameters}
\usage{
font_details_pt(gp = gpar())
}
\arguments{
\item{gp}{Graphical parameters}
}
\description{
Calculates generic font height and descent from given graphical parameters
}
