% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridtext.R
\docType{package}
\name{gridtext}
\alias{gridtext}
\title{Improved text rendering support for grid graphics}
\description{
The gridtext package provides two new grobs, \code{\link[=richtext_grob]{richtext_grob()}} and
\code{\link[=textbox_grob]{textbox_grob()}}, which support drawing of formatted text labels and
formatted text boxes, respectively.
}
