% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob-descent.R
\name{grob_descent}
\alias{grob_descent}
\alias{grob_descent_pt}
\title{Calculate descent and other text properties, either from grob or from graphical
parameters}
\usage{
grob_descent(grob)

grob_descent_pt(grob)
}
\arguments{
\item{grob}{The grob for which we want information}
}
\description{
Calculate descent and other text properties, either from grob or from graphical
parameters
}
