% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob-width-pt.R
\name{grob_width_pt}
\alias{grob_width_pt}
\alias{grob_height_pt}
\title{Get width and height info of a grob, in pt}
\usage{
grob_width_pt(grob)

grob_height_pt(grob)
}
\arguments{
\item{grob}{The grob for which we want the width or height}
}
\description{
Get width and height info of a grob, in pt
}
