% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels-grob.R
\name{label_grob}
\alias{label_grob}
\title{create individual labels for labels_grob}
\usage{
label_grob(label, x = unit(0.5, "npc"), y = unit(0.5, "npc"), hjust = 0.5,
  vjust = 0.5, padding = margin(0, 0, 0, 0), margin = margin(0, 0, 0, 0),
  angle = 0, gp = gpar(), debug = TRUE)
}
\description{
create individual labels for labels_grob
}
