% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels-grob.R
\name{labels_grob}
\alias{labels_grob}
\title{Create grob representing one or more labels from tibble}
\usage{
labels_grob(label_data, gp = gpar(), debug = TRUE)
}
\arguments{
\item{label_data}{Tibble holding the label data. At a minimum, needs a
\code{label} column holding the text labels to be drawn}

\item{gp}{Additional graphical parameters not provided via \code{label_data}.}

\item{debug}{Bool indicating whether debugging info should be drawn.}
}
\description{
Create grob representing one or more labels from tibble
}
\examples{
library(grid)
library(tibble)

label_data <- tibble(
  label = c("Descenders: pgqjy", "This is a label\\nwith two lines", "Hello!"),
  x = unit(.4, "npc"),
  y = unit(c(.9, .5, .3), "npc"),
  hjust = 0,
  vjust = 1,
  angle = c(0, 45, -45),
  padding = list(margin(2, 0, 2, 0)),
  margin = list(margin(5, 5, 5, 5))
)

grid.newpage()
grid.draw(labels_grob(label_data, gp = gpar(col = "red"), debug = TRUE))
}
