% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels-grob.R
\name{make_box_grob}
\alias{make_box_grob}
\title{make frame around text grobs}
\usage{
make_box_grob(grob, width_pt, height_pt, descent_pt, x = unit(0.5,
  "npc"), y = unit(0.5, "npc"), hjust = 0.5, vjust = 0.5,
  padding = list(mar(0, 0, 0, 0)), margin = list(mar(0, 0, 0, 0)),
  angle = 0, fill = NA, color = NA, frame_color = NULL, ...,
  debug = FALSE)
}
\description{
make frame around text grobs
}
