% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels-grob.R
\name{make_label_grob}
\alias{make_label_grob}
\title{create individual labels for labels_grob}
\usage{
make_label_grob(label, hjust_int = 0.5, vjust_int = 0.5,
  fontfamily = "", fontface = "plain", fontsize = 12,
  lineheight = 1.1, color = "black", ...)
}
\description{
create individual labels for labels_grob
}
