% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_gridtext_options}
\alias{set_gridtext_options}
\title{Sets the default options for the gridtext  package.}
\usage{
set_gridtext_options(overwrite = FALSE)
}
\arguments{
\item{overwrite}{Overwrite all options with gridtext default settings}
}
\description{
Options:
gridtext.stop_if_tag_is_unimplemented with default behaviour: Stop if an unimplemented tag is incountered.
}
\examples{
options(gridtext.stop_if_tag_is_unimplemented = FALSE)
}
