% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-box-grob.R
\name{text_box_grob}
\alias{text_box_grob}
\title{Text box grob}
\usage{
text_box_grob(text, width = NULL, height = NULL, x = unit(0.5,
  "npc"), y = unit(0.5, "npc"), hjust = 0.5, vjust = 0.5,
  gp = gpar(), use_markdown = TRUE)
}
\arguments{
\item{text}{Character vector containing markdown/html string to draw.}

\item{width, height}{Unit objects specifying width and height of the
grob; height is currently ignored.}

\item{x, y}{Unit objects specifying the location of the reference point.}

\item{hjust, vjust}{Numerical values specifying the location of the grob
relative to the reference point.}

\item{gp}{Other graphical parameters for drawing}

\item{use_markdown}{Should the \code{text} input be treated as markdown?}
}
\description{
Text box grob
}
\examples{
library(grid)
grid.newpage()
grid.draw(text_box_grob("Some text **in bold.**"))
}
