% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-details.R
\name{text_details}
\alias{text_details}
\title{Calculate text details for a given text label}
\usage{
text_details(label, gp = gpar())
}
\arguments{
\item{label}{Character vector containing the label. Can handle only one label at a time.}

\item{gp}{Grid graphical parameters defining the font (\code{fontfamily}, \code{fontface}, and
\code{fontface} should be defined).}
}
\description{
Calculate text details for a given text label
}
\examples{
text_details("Hello world!", gpar(fontfamily = "", fontface = "plain", fontsize = 12))
text_details("Hello world!", gpar(fontfamily = "", fontface = "plain", fontsize = 24))
text_details(
  "Hello world\\nwith newline",
  gpar(fontfamily = "", fontface = "plain", fontsize = 12)
)
}
