% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box-collection-grob.R
\name{box_collection_grob}
\alias{box_collection_grob}
\title{Draw a collection of box grobs}
\usage{
box_collection_grob(..., width = NULL, height = NULL, vp = NULL,
  debug = FALSE)
}
\arguments{
\item{...}{One or more box grobs.}

\item{width, height}{Width and height of the box collection grob. If either is set
to \code{NULL} (the default), it is calculated from the box grobs.}

\item{vp}{Grid viewport.}

\item{debug}{Bool indicating whether debug info should be drawn or not.}
}
\description{
Draw a collection of box grobs.
}
\examples{
# no examples yet
}
