% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-rich-text.R
\name{draw_rich_text}
\alias{draw_rich_text}
\title{Draw rich text}
\usage{
draw_rich_text(contents, hjust = 0.5, x_pt = 50, y_pt = 100)
}
\arguments{
\item{contents}{character vector containing html string to draw}

\item{hjust}{horizontal justification}

\item{x_pt}{x location, in points}

\item{y_pt}{y location, in points}
}
\description{
Draw rich text
}
\examples{
library(grid)
grid.newpage()
draw_rich_text("This is the first line.<br>This is the second line.<br>And some <b>text in bold.</b>")
}
