% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels-grob.R
\name{labels_grob}
\alias{labels_grob}
\title{Create grob representing one or more labels}
\usage{
labels_grob(label_data, align_frames = FALSE, debug = FALSE)
}
\arguments{
\item{label_data}{Tibble holding the label data. At a minimum, needs a
\code{label} column holding the text labels to be drawn}

\item{align_frames}{Bool indicating whether frames should have equal
sizes (\code{TRUE}) or not (\code{FALSE}).}

\item{debug}{Bool indicating whether debugging info should be drawn.}

\item{gp}{Additional graphical parameters not provided via \code{label_data}.}
}
\description{
The function \code{labels_grob()} creates one or more labels. All label data is provided
in the form of a data frame/tibble holding aesthetic values as columns and label
information as rows. The padding and margin information needs to be provided in a
list column.
}
\examples{
library(grid)
library(tibble)

label_data <- tibble(
  label = c("Descenders: pgqjy", "This is a label\\nwith two lines", "Hello!"),
  x = unit(.4, "npc"),
  y = unit(c(.9, .5, .3), "npc"),
  hjust = 0,
  vjust = 0.5,
  hjust_int = 1, # internal hjust, vjust, defines how text label
  vjust_int = 1, # is positioned inside the box
  color = "blue",
  fill = "azure1",
  fontsize = 10,
  fontfamily = "Comic Sans MS",
  angle = c(0, 45, -45),
  padding = list(mar(5, 5, 3, 5)),
  margin = list(mar(5, 5, 5, 5))
)

grid.newpage()
g <- labels_grob(label_data, align_frames = TRUE)
grid.draw(g)


label_data <- tibble(
  label = c("This", "is", "a", "test", "x-axis"),
  x = unit(c(.1, .3, .5, .7, .9), "npc"),
  y = unit(0.5, "npc"),
  hjust = 0.5,
  vjust = 1,
  hjust_int = 0.5,
  vjust_int = 1,
  angle = 0,
  fontsize = 10, fontfamily = "Comic Sans MS",
  padding = list(mar(5, 5, 3, 5)),
  margin = list(mar(5, 5, 5, 5))
)
grid.newpage()
g <- labels_grob(label_data, align_frames = TRUE, debug = TRUE)
grid.draw(g)


label_data <- tibble(label = "abcqgy", fontsize = 50, fontfamily = "Comic Sans MS")
grid.newpage()
grid.draw(labels_grob(label_data))
}
