% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels-grob.R
\name{margin}
\alias{margin}
\title{Like the margin function in ggplot2}
\usage{
margin(t = 0, r = 0, b = 0, l = 0, unit = "pt")
}
\arguments{
\item{t, r, b, l}{top, right, bottom, left margins}

\item{unit}{unit in which margins are measured}
}
\description{
Like the margin function in ggplot2
}
