% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rich-text-grob.R
\name{rich_text_grob}
\alias{rich_text_grob}
\title{Rich-text grob}
\usage{
rich_text_grob(text, x = unit(0.5, "npc"), y = unit(0.5, "npc"),
  hjust = 0.5, vjust = 0.5, rot = 0, default.units = "npc",
  name = NULL, gp = gpar(), vp = NULL, use_markdown = TRUE)
}
\arguments{
\item{text}{Character vector containing markdown/html string to draw.}

\item{x, y}{Unit objects specifying the location of the reference point.}

\item{hjust, vjust}{Numerical values specifying the location of the grob
relative to the reference point.}

\item{rot}{Angle of rotation for text, in degrees.}

\item{default.units}{Units of \code{x} and \code{y} if these are provided only as
numerical values.}

\item{name}{Name of the grob.}

\item{gp}{Other graphical parameters for drawing.}

\item{vp}{Viewport.}

\item{use_markdown}{Should the \code{text} input be treated as markdown?}
}
\description{
Rich-text grob
}
\examples{
library(grid)

text <- c(
  "Some text **in bold.**", "Linebreaks<br>Linebreaks<br>Linebreaks",
  "*x*<sup>2</sup> + 5*x* + *C*<sub>i</sub>",
  "Mixing **it** all *up,<br> for good measure.*"
)

x <- c(.2, .1, .7, .9)
y <- c(.8, .4, .2, .6)
rot <- c(0, 0, 45, -45)
gp = gpar(col = c("black", "red"), fontfamily = c("Palatino", "Courier", "Times", "Helvetica"))
hjust <- c(0.5, 0, 0, 1)
vjust <- c(0.5, 1, 0, 0.5)

grid.newpage()
g <- rich_text_grob(text, x, y, hjust = hjust, vjust = vjust, rot = rot, gp = gp)
grid.draw(g)
grid.points(x, y, default.units = "npc")
}
