% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-rich-text.R
\name{rich_text_grob}
\alias{rich_text_grob}
\title{Rich-text grob}
\usage{
rich_text_grob(contents, x = unit(0.5, "npc"), y = unit(0.5, "npc"),
  hjust = 0, angle = 0)
}
\arguments{
\item{contents}{character vector containing html string}

\item{x}{x location}

\item{y}{y location}

\item{hjust}{horizontal justification}
}
\description{
Rich-text grob
}
\examples{
library(grid)
grid.newpage()
grid.draw(rich_text_grob("Some text <b>in bold.</b>"))
}
