% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-rich-text.R
\name{rich_text_grob}
\alias{rich_text_grob}
\alias{markdown_grob}
\title{Rich-text grob}
\usage{
rich_text_grob(contents, ..., hjust_int = 0)

markdown_grob(contents, ...)
}
\arguments{
\item{contents}{character vector containing html string}

\item{...}{Other arguments handed off to \code{\link[=box_grob]{box_grob()}}}

\item{hjust_int}{Internal horizontal justification (0 = left, 0.5 = center, 1 = right)}
}
\description{
Rich-text grob
}
\examples{
library(grid)
grid.newpage()
grid.draw(rich_text_grob("Some text <b>in bold.</b>"))
}
