% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-rich-text.R
\name{rich_text_grob}
\alias{rich_text_grob}
\alias{markdown_grob}
\title{Rich-text grob}
\usage{
rich_text_grob(contents, ..., width_pt = 300, gp = NULL)

markdown_grob(contents, ...)
}
\arguments{
\item{contents}{character vector containing html string}

\item{...}{Other arguments handed off to \code{\link[=box_grob]{box_grob()}}}

\item{width_pt}{width, in points}

\item{gp}{Other graphical parameters for drawing}
}
\description{
Rich-text grob
}
\examples{
#library(grid)
#grid.newpage()
#grid.draw(rich_text_grob("Some text <b>in bold.</b>"))
}
