% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels-grob.R
\name{text_grob}
\alias{text_grob}
\title{create individual labels for labels_grob}
\usage{
text_grob(label, hjust = 0.5, vjust = 0.5, fontfamily = "",
  fontface = "plain", fontsize = 12, lineheight = 1.1,
  color = "black", ...)
}
\description{
create individual labels for labels_grob
}
